moved core_models to common_model
This change is needed to keep the modular behavior of different views
