+++ app/hindsite/home/templates/facilitator-home.html
@@ -19,7 +19,10 @@
                         aria-label="Slide 2"></button>
                 </div>
                 <div class="carousel-inner h-100">
+                    <!-- Load board -->
                     <div class="carousel-item active">
+                            <!-- load fields -->
+                                <!-- load cards -->
                         <div class="card">
                             <div class="card-body">
                                 <p class="card-text">
